Move function definition parser
